# swagger_Timebank_v0.4.yaml
openapi: 3.0.1
info:
  title: Timebank
  description: 'This is API description for Timebank project.'
  version: 0.4.0
servers:
- url: http://localhost:5000/api/v1
...
paths:
...
  /services:
    get:
      security: []
      tags:
      - service
      summary: Get all services
      description: Get all services in the Timebank
      parameters:
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/sortFieldParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /service/{id}:
    parameters:
      - name: id
        in: path
        description: ID of service to use
        required: true
        schema:
          type: integer
          format: int64
    get:
      security: []
      tags:
      - service
      summary: Find service by ID
      operationId: getServiceById
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
    put:
      security:
      - bearerAuth: []
      tags:
      - service
      summary: Updates a service
      operationId: updateService
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
    delete:
      security:
      - bearerAuth: []
      tags:
      - service
      summary: Deletes a service
      operationId: deleteService
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /service/{title}:
    parameters:
      - name: title
        in: path
        description: ID of service to use
        required: true
        schema:
          type: integer
          format: int64
    get:
      security: []
      tags:
      - service
      summary: Find service by ID
      operationId: getServiceByTitle
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /service/{user_id}:
    parameters:
      - name: user_id
        in: path
        description: ID of service to use
        required: true
        schema:
          type: integer
          format: int64
    get:
      security: []
      tags:
      - service
      summary: Find service by ID
      operationId: getServiceByUserID
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'  
  /service-create:
    post:
      security:
      - bearerAuth: []
      tags:
        - service
      summary: Create a service in the database
      operationId: createService
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Service'
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Service:
      required:
        - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Pokosim travnik
        estimate:
          type: integer
          format: int64
          example: 2
    ServiceUpdate:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Umyjem okna na mrakodrape
        estimate:
          type: integer
          format: int64
          example: 2
...
  parameters:
    sortOrderParam:
      in: query
      name: sort
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
      description: Used to sort return values in ascending or descending order
    sortFieldParam:
      in: query
      name: field
      required: false
      schema:
        type: string
      description: Used to select sorting field from the model
...
  responses:
    NotFound:
      description: No entries found.
    NotAuthorized:
      description: Authorization information is missing or invalid.
    BadParam:
      description: Bad paramters.
    ServerError:
      description: Unexpected response error. Server fault.
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: bearer
      bearerFormat: JWT